% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{DCA}
\alias{DCA}
\title{This function computes the Discriminant Component Analysis (RCA)
algorithm.}
\usage{
DCA(X, chunklets, D, total_dims = NULL)
}
\arguments{
\item{X}{Input numeric matrix where each row is a data point and each
column is a variable.}

\item{chunklets}{An integer vector indicating to which chunklet each data
point belongs to. If a data point does not belong to
a chunklet, then its chunklet value should be set to -1
otherwise all chunklets should be represented by
a non-negative integer. All data points in the same chunklet
must have the same class label.}

\item{D}{A 0-1 symmetric matrix where a 1 at \code{D[i,j]} indicates that
chunklets \code{i} and \code{j} share a negative constraint i.e chunklets
\code{i} and \code{j} do not belong to the same class. The position of the
chunklets in the rows and columns of \code{D} is in
increasing order of the integers representing the chunklets. For
example, if we have chunklet 1 and chunklet 2, then the first row
and first column in \code{D} will reference chunklet 1, and
the second row and second column in \code{D} will reference
chunklet 2.}

\item{total_dims}{The number of dimensions on which the data should be projected
on after the DCA metric has been learned. Default value is set
to \code{total_dims = NULL} in which case the value
\code{total_dims = ncol(X)} is used as default.}
}
\value{
This function returns a list with the following items:
\item{DCATransform}{The transformation matrix.}
\item{TransformedX}{The transformed original data \code{X} which was transformed
                    using the DCA trasformation matrix i.e \eqn{TransformedX = X *
                    DCATransform}}.
}
\description{
This function computes the Discriminant Component Analaysis
             (DCA) algorithm as described in [1]. See the Vignette
             by using the command \code{browseVignette("DistanceLearning")}
             for an introduction to using DCA.
}
\details{
See the Vignette by using the command
         \code{browseVignette("DistanceLearning")}
         for an introduction to using DCA.
}
\examples{
# Load data from package DistanceLearning
library(DistanceLearning)
library(class)
fname <- system.file("extdata", "example_data.csv", package="DistanceLearning")
df <- read.csv(fname)
Y <- as.integer(df$y)
X <- as.matrix(df[,c(2,3)])
sample_points <- sample(1:nrow(X), 180, replace = FALSE)
subX <- X[sample_points,]
subY <- Y[sample_points]

D <- matrix(c(0,1,1,0), nrow = 2, ncol = 2)

# Predict class labels for newX and get the metric
result <- DCA(subX, chunklets = subY, D = D)
DCAMetric <- result$DCATransform
transformednewX <- result$transformedX

# Get the accuracy of KNN classification without applying the new metric
yhat <- knn(subX, X[-sample_points,], subY, k = 5)
Accuracy <- length(which(Y[-sample_points] == yhat))/length(Y[-sample_points])
Accuracy

# Get the accuracy of KNN classification after applying the new metric
transformedX <- X[-sample_points,] \%*\% DCAMetric
yhat2 <- knn(transformednewX, transformedX, subY, k = 5)
Accuracy2 <- length(which(Y[-sample_points] == yhat2))/length(Y[-sample_points])
Accuracy2

}
\references{
[1] Hoi, S.C.H., Liu, W., Lyu, M.R., Ma, W.Y.: Learning
                distance metrics with contextual constraints for image
                retrieval. In: IEEE Conference on Computer Vision and
                Pattern Recognition.
}
\seealso{
\code{\link{RCA}}
}
\author{
Carl Tony Fakhry, Ping Chen, Rahul Kulkarni and Kourosh Zarringhalam
}
