\name{DistanceLearning-package}
\alias{DistanceLearning-package}
\docType{package}
\title{

An R package that implements 11 distance metric learning methods

}
\description{

This package implements 11 distance metric
       learning methods. The methods are:
       Free Energy Nearest Neighbor (FENN),
       Free Energy Nearest Neighbor with dimension reduction (subFENN),
       Discriminant Adaptive Nearest Neighbor (DANN),
       iterative Discriminant Adaptive Nearest Neighbor (iDANN),
       Locally Adaptive Metric Nearest Neighbor (ADAMENN),
       Locally Adaptive Metric Nearest Neighbor (iADAMENN),
       Xing's method,
       Relevant Component Analaysis (RCA),
       Linear Fisher Discriminant Analaysis (LFDA),
       Discriminant Component Analaysis (DCA),
       Neighborhood Component Analysis (NCA).
       All methods are implemented using RcppArmadillo in order to attain improved performance.
       A vignette detailing the usage of each method is attached to the package.

}
\details{
\tabular{ll}{
Package: \tab DistanceLearning\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-07-25\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
Carl Tony Fakhry, Ping Chen, Rahul Kulkarni and Kourosh Zarringhalam

Maintainer: Carl Tony Fakhry <cfakhry@cs.umb.edu>
}
\references{
Carl Tony Fakhry, Ping Chen, Rahul Kulkarni and Kourosh Zarringhalam. "A Free Energy Based Approach for Distance Metric Learning", Submitted.
}
